app/requirements.txt
flask==3.1.0
pandas==2.2.3
azure-storage-blob==12.24.0
azure-identity==1.21.0
